fix: more styles chanes
