+++ tldraw/apps/tldraw-logseq/src/components/Button/CircleButton.tsx
@@ -1,0 +1,41 @@
+import React from 'react'
+import { TablerIcon } from '../icons'
+
+export const CircleButton = ({
+  active,
+  style,
+  icon,
+  otherIcon,
+  onClick,
+}: {
+  active?: boolean
+  style?: React.CSSProperties
+  icon: string
+  otherIcon?: string
+  onClick: () => void
+}) => {
+  const [recentlyChanged, setRecentlyChanged] = React.useState(false)
+
+  React.useEffect(() => {
+    setRecentlyChanged(true)
+    const timer = setTimeout(() => {
+      setRecentlyChanged(false)
+    }, 500)
+    return () => clearTimeout(timer)
+  }, [active])
+
+  return (
+    <button
+      data-active={active}
+      data-recently-changed={recentlyChanged}
+      style={style}
+      className="tl-circle-button"
+      onMouseDown={onClick}
+    >
+      <div className="tl-circle-button-icons-wrapper" data-icons-count={otherIcon ? 2 : 1}>
+        {otherIcon && <TablerIcon name={otherIcon} />}
+        <TablerIcon name={icon} />
+      </div>
+    </button>
+  )
+}
