+++ tldraw/apps/tldraw-logseq/src/components/QuickSearch/index.ts
@@ -1,0 +1,1 @@
+export * from './QuickSearch'
